+++ packages/admin/resources/views/livewire/components/settings/attributes/index.blade.php
@@ -1,40 +1,31 @@
 <div class="flex-col space-y-4">
-  <div class="text-right">
-    <x-hub::button tag="a" href="{{ route('hub.attributes.create') }}">Create Attribute</x-hub::button>
-  </div>
   <x-hub::table>
     <x-slot name="head">
       <x-hub::table.heading>
-        {{ __('adminhub::global.name') }}
-      </x-hub::table.heading>
-      <x-hub::table.heading sortable>
         {{ __('adminhub::global.attribute_type') }}
       </x-hub::table.heading>
       <x-hub::table.heading sortable>
-        {{ __('adminhub::global.type') }}
+        {{ __('adminhub::global.attribute_groups') }}
       </x-hub::table.heading>
       <x-hub::table.heading>
-        {{ __('adminhub::global.required') }}
+        {{ __('adminhub::global.attributes') }}
       </x-hub::table.heading>
       <x-hub::table.heading></x-hub::table.heading>
     </x-slot>
     <x-slot name="body">
-      @forelse($attributes as $attribute)
+      @forelse($this->attributeTypes as $handle => $attributeType)
         <x-hub::table.row>
           <x-hub::table.cell>
-            {{ $attribute->name['en'] }}
+            {{ class_basename($attributeType) }}
           </x-hub::table.cell>
           <x-hub::table.cell>
-            {{ class_basename($attribute->attribute_type) }}
+            {{ $this->getStats($attributeType)['group_count'] }}
           </x-hub::table.cell>
           <x-hub::table.cell>
-            {{ class_basename($attribute->type) }}
-          </x-hub::table.cell>
-          <x-hub::table.cell>
-            <x-hub::icon :ref="$attribute->required ? 'check' : 'x'" :class="$attribute->required ? 'text-green-500' : 'text-red-500'" style="solid" />
+            {{ $this->getStats($attributeType)['attribute_count'] }}
           </x-hub::table.cell>
           <x-hub::table.cell class="text-right">
-            <a href="{{ route('hub.attributes.show', $attribute) }}" class="text-indigo-500 hover:underline">
+            <a href="{{ route('hub.attributes.show', $handle) }}" class="text-indigo-500 hover:underline">
                 {{ __('adminhub::settings.currencies.index.table_row_action_text') }}
               </a>
           </x-hub::table.cell>
@@ -46,7 +37,4 @@
       @endforelse
     </x-slot>
   </x-hub::table>
-  <div>
-    {{ $attributes->links() }}
-  </div>
 </div>