+++ packages/admin/src/Http/Livewire/Components/Settings/Attributes/AttributesIndex.php
@@ -2,7 +2,9 @@
 
 namespace GetCandy\Hub\Http\Livewire\Components\Settings\Attributes;
 
+use GetCandy\Facades\AttributeManifest;
 use GetCandy\Models\Attribute;
+use GetCandy\Models\AttributeGroup;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -11,14 +13,43 @@ class AttributesIndex extends Component
     use WithPagination;
 
     /**
+     * Return the attribute types computed property.
+     *
+     * @return void
+     */
+    public function getAttributeTypesProperty()
+    {
+        return AttributeManifest::getTypes();
+    }
+
+    /**
+     * Returns stats for an attribute type.
+     *
+     * @param string $attributeType
+     *
+     * @return int
+     */
+    public function getStats($attributeType)
+    {
+        $groups = AttributeGroup::whereAttributableType($attributeType)->get();
+
+        return [
+            'group_count'     => $groups->count(),
+            'attribute_count' => Attribute::whereIn(
+                'attribute_group_id',
+                $groups->pluck('id')->toArray()
+            )->count(),
+        ];
+    }
+
+    /**
      * Render the livewire component.
      *
      * @return \Illuminate\View\View
      */
     public function render()
     {
-        return view('adminhub::livewire.components.settings.attributes.index', [
-            'attributes' => Attribute::paginate(5),
-        ])->layout('adminhub::layouts.base');
+        return view('adminhub::livewire.components.settings.attributes.index')
+            ->layout('adminhub::layouts.base');
     }
 }