+++ packages/core/database/migrations/2022_01_12_100000_add_columns_to_attributes_table.php
@@ -1,0 +1,26 @@
+<?php
+
+use GetCandy\Base\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddColumnsToAttributesTable extends Migration
+{
+    public function up()
+    {
+        Schema::table($this->prefix.'attributes', function (Blueprint $table) {
+            $table->boolean('searchable')->after('system')->default(true)->index();
+            $table->boolean('filterable')->after('system')->default(false)->index();
+            $table->string('validation_rules')->after('system')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        Schema::table($this->prefix.'attributes', function (Blueprint $table) {
+            $table->dropColumn('searchable');
+            $table->dropColumn('filterable');
+            $table->dropColumn('validation_rules');
+        });
+    }
+}