+++ packages/core/database/migrations/2022_01_12_100002_add_default_column_to_tax_classes_table.php
@@ -1,0 +1,22 @@
+<?php
+
+use GetCandy\Base\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddDefaultColumnToTaxClassesTable extends Migration
+{
+    public function up()
+    {
+        Schema::table($this->prefix.'tax_classes', function (Blueprint $table) {
+            $table->boolean('default')->after('name')->index()->default(false);
+        });
+    }
+
+    public function down()
+    {
+        Schema::table($this->prefix.'tax_classes', function (Blueprint $table) {
+            $table->dropColumn('default');
+        });
+    }
+}