+++ packages/core/src/Facades/FieldTypeManifest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace GetCandy\Facades;
+
+use GetCandy\Base\FieldTypeManifestInterface;
+use Illuminate\Support\Facades\Facade;
+
+class FieldTypeManifest extends Facade
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected static function getFacadeAccessor()
+    {
+        return FieldTypeManifestInterface::class;
+    }
+}